requirements.txt
# Core ML
torch>=2.1
transformers>=4.45         # required for V-JEPA 2 (added in 4.45+)
accelerate>=0.30
peft>=0.10                 # LoRA / QLoRA

# Quantization (for QLoRA only — comment out if not running QLoRA)
bitsandbytes>=0.43

# Video loading
decord>=0.6
# Alternative on macOS: eva-decord (uncomment if needed)
# eva-decord>=0.6.1

# Data
pandas>=2.0
numpy>=1.24
Pillow>=10.0

# Logging / utility
wandb>=0.16
tqdm>=4.66
PyYAML>=6.0
